+++ packages/api/src/resolvers/queries/events/eventStatistics.ts
@@ -1,13 +1,15 @@
 import { log } from '@unchainedshop/logger';
 import { Context } from '../../../context.js';
+import { DateFilterInput } from '@unchainedshop/utils';
 
 export default async function eventStatistics(
   root: never,
-  params: { types?: string[]; from?: Date; to?: Date },
+  params: { types?: string[]; dateRange?: DateFilterInput },
   { userId, modules }: Context,
 ) {
-  log(`query eventStatistics ${params.types || ''} ${params.from || ''} ${params.to || ''}`, {
+  log(`query eventStatistics ${(params?.types || []).join(', ')}`, {
     userId,
+    ...(params?.dateRange || {}),
   });
 
   return modules.events.getReport(params);