+++ packages/api/src/resolvers/queries/worker/workStatistics.ts
@@ -1,13 +1,15 @@
 import { log } from '@unchainedshop/logger';
 import { Context } from '../../../context.js';
+import { DateFilterInput } from '@unchainedshop/utils';
 
 export default async function workStatistics(
   root: never,
-  params: { types?: string[]; from?: Date; to?: Date },
+  params: { types?: string[]; dateRange?: DateFilterInput },
   { modules, userId }: Context,
 ) {
-  log(`query workStatistics ${params.from || ''} ${params.to || ''}`, {
+  log(`query workStatistics ${(params?.types || []).join(', ')}`, {
     userId,
+    ...(params.dateRange || {}),
   });
 
   return modules.worker.getReport(params);