+++ packages/core-worker/src/module/configureWorkerModule.ts
@@ -9,7 +9,12 @@ import {
   ModuleInput,
 } from '@unchainedshop/mongodb';
 import { emit, registerEvents } from '@unchainedshop/events';
-import { buildObfuscatedFieldsFilter, SortDirection, SortOption } from '@unchainedshop/utils';
+import {
+  buildObfuscatedFieldsFilter,
+  DateFilterInput,
+  SortDirection,
+  SortOption,
+} from '@unchainedshop/utils';
 import { WorkQueueCollection } from '../db/WorkQueueCollection.js';
 import { DIRECTOR_MARKED_FAILED_ERROR, WorkerDirector } from '../director/WorkerDirector.js';
 import { WorkerEventTypes } from '../director/WorkerEventTypes.js';
@@ -90,7 +95,7 @@ export type WorkerModule = {
     referenceDate: Date;
   }) => Promise<Array<Work>>;
 
-  getReport: (params?: { types?: string[]; from?: Date; to?: Date }) => Promise<WorkerReport[]>;
+  getReport: (params: { types?: string[]; dateRange?: DateFilterInput }) => Promise<WorkerReport[]>;
 };
 
 const WORK_STATUS_FILTER_MAP = {
@@ -642,22 +647,22 @@ export const configureWorkerModule = async ({
       );
     },
 
-    getReport: async ({ types, from, to } = { types: null, from: null, to: null }) => {
+    getReport: async ({ types, dateRange } = { types: null, dateRange: {} }) => {
       const pipeline = [];
       const matchConditions = [];
       // build date filter based on provided values it can be a range if both to and from is supplied
       // a upper or lowe limit if either from or to is provided
       // or all if none is provided
-      if (from || to) {
+      if (dateRange?.start || dateRange?.end) {
         const dateConditions = [];
-        if (from) {
-          const fromDate = new Date(from);
+        if (dateRange?.start) {
+          const fromDate = new Date(dateRange?.start);
           dateConditions.push({
             $or: [{ created: { $gte: fromDate } }, { updated: { $gte: fromDate } }],
           });
         }
-        if (to) {
-          const toDate = new Date(to);
+        if (dateRange?.end) {
+          const toDate = new Date(dateRange?.end);
           dateConditions.push({
             $or: [{ created: { $lte: toDate } }, { updated: { $lte: toDate } }],
           });